chore: Add pre.json for changeset pre-release.
chore: Initialize pre-release mode with `changesets`

This commit initializes the pre-release mode using `changesets`. It creates a `pre.json` file within the `.changeset` directory to configure the pre-release process. This file specifies the `mode` as "pre" and the `tag` as "rc", indicating that release candidates will be used for pre-release versions.

Furthermore, the `initialVersions` field sets the initial version for the "tempeh" package to "5.0.2" as the starting point for the pre-release series. This will ensure that subsequent pre-release versions for "tempeh" build upon this baseline. The `changesets` array is initialized as empty, signifying that no changesets are currently associated with the pre-release. This setup prepares the repository for managing and publishing pre-release versions of the "tempeh" package using `changesets`.